requirements.txt
torch>=2.2.0
transformers>=4.41.0
peft>=0.11.0
datasets>=2.19.0
pillow>=10.3.0
numpy>=1.26.0
matplotlib>=3.8.0

